requirements.txt
pypdf
einops 
torch
accelerate 
deepspeed
accelerate
transformers 
langchain 
langchain-community
bitsandbytes
sentence_transformers
llama-index
llama-index-embeddings-langchain
llama-index-embeddings-google
llama-index-embeddings-huggingface 
llama-index-embeddings-gemini
llama-index-llms-groq
llama-index-llms-huggingface
llama-index-llms-langchain
google-generativeai
nest_asyncio
fastapi